Wrote new MixtureModel constructor method; updated README file.
Wrote a new MixtureModel constructor that takes a MixtureModel as an argument. The constructor then tries to build a new Mixture model that is the target (and argument) Hellinger distance away from the argument MixtureModel.

Updated the README file to reflect that the project now contains two data stream generators.