Update README; Adjust method for coming up with new weights and means for the post-"concept drift" mixture model.
Added a paragraph to the README to acknowledge the Apache Commons Mathematics Library as a dependency.

Adjusted MixtureModel's second constructor method for developing a new MixtureModel that is the desired Hellinger distance away from the argument MixtureModel. The new random number is now scaled by (1.0 - Math.pow(targetDist, 2.0)) instead of by only (1.0 - targetDist). The goal of this is to make it easier to generate mixture models that are a small Hellinger distance apart (i.e. H < 0.4).