Rewrote second MixtureModel constructor method.
Rewrote MixtureModel's second constructor which takes another mixture model and a target Hellinger distance as additional arguments.
The goal is to generate a new mixture model which is the target Hellinger distance away from the argument mixture model.
This is done by generating a completely random mixture model and then adjusting its mean points and weight array back towards the original mixture model. The weighting used is 's', equal to the square of the target distance, for the newly generated mixture model's values and '1 - s' for the original mixture model.
This process is more likely to work for larger target Hellinger distances than smaller ones and should be revisited to be optimized.